+++ valpas-web/src/components/forms/InputContainer.tsx
@@ -1,5 +1,6 @@
 import bem from "bem-ts"
 import React from "react"
+import { joinClassNames } from "../../utils/classnames"
 
 export type InputContainerProps = {
   bemBase: string
@@ -7,15 +8,17 @@ export type InputContainerProps = {
   label?: React.ReactNode
   icon?: React.ReactNode
   error?: React.ReactNode
+  className?: string
 }
 
 export const InputContainer = (props: InputContainerProps) => {
   const b = bem(props.bemBase)
   return (
     <label
-      className={b({
-        withicon: Boolean(props.icon),
-      })}
+      className={joinClassNames(
+        b({ withicon: Boolean(props.icon) }),
+        props.className
+      )}
     >
       {props.label && <div className={b("label")}>{props.label}</div>}
       <div className={b("inputcontainer")}>
