+++ valpas-web/src/views/VirkailijaApp.tsx
@@ -4,7 +4,6 @@ import { fetchYlatasonOrganisaatiotJaKayttooikeusroolit } from "../api/api"
 import { useApiOnce } from "../api/apiHooks"
 import { isSuccess } from "../api/apiUtils"
 import { Page } from "../components/containers/Page"
-import { MainNavigation } from "../components/navigation/MainNavigation"
 import { t } from "../i18n/i18n"
 import {
   CurrentUser,
@@ -44,12 +43,6 @@ type VirkailijaRoutesProps = {
 
 const VirkailijaRoutes = ({ user }: VirkailijaRoutesProps) => {
   const basePath = useBasePath()
-  const navOptions = [
-    {
-      key: "hakutilanne",
-      display: t("ylänavi__hakutilanne"),
-    },
-  ]
 
   const organisaatiotJaKayttooikeusroolit = useApiOnce(
     fetchYlatasonOrganisaatiotJaKayttooikeusroolit
@@ -66,11 +59,6 @@ const VirkailijaRoutes = ({ user }: VirkailijaRoutesProps) => {
       </Route>
       {isFeatureFlagEnabled() && (
         <Route exact path={`${basePath}/oppijat`}>
-          <MainNavigation
-            selected="hakutilanne"
-            options={navOptions}
-            onChange={() => null}
-          />
           <PerusopetusView
             kayttooikeusroolit={organisaatiotJaKayttooikeusroolit.data}
           />
