+++ valpas-web/src/views/hakutilanne/PerusopetusView.less
@@ -1,0 +1,4 @@
+.perusopetusview__organisaatiovalitsin {
+  margin-bottom: 2rem;
+  padding-left: 0.5rem;
+}
